requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
python-dateutil==2.8.2
email-validator==2.1.0
